Revert Violations changes
